requirements.txt
Django>=5.0
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1

Pillow>=10.0
python-dotenv>=1.0

django-allauth>=0.61
django-crispy-forms>=2.1

requests>=2.31